refactor: extract runtime artifact builders
